bug_finder_adk/requirements.txt
google-adk
google-agents
google-generativeai
rich
streamlit  # Optional: Uncomment if you want to use Streamlit UI 